The below changes were for patching our fonts to disable the smcp feature.
Add the -v switch to specify a version number for the version string. No suffix is added to the version string if this switch is used.

Modify so that if only default settings are applied, the font name and version do not have suffixes added to them. This allows cmd "scripts" to be applied to a font which don't change the font name or version.

git-svn-id: file:///home/mhosken/Work/MSEA/Personal/TimEves/subversion/repositories/typetuner/trunk@108 4150712b-26e8-44b9-9db8-fb9803e0bbe7