Fix issue where git-sync updates would wipe out the apps view of the storage
Because the app mounted a subPath of the volume, the immediate child of which was a symbolic link managed by git-sync, this link would be resolved only once at container start up and not re-resolved after if git-sync changed worktrees.  Eventually the old worktree would be reaped leading to no tunable fonts at all.

This fixes that by forcing path resolution to happen on every page load instead of once,